refactor(web): share file uploads outside the matter route
